Update My Books page
This takes care of a few tasks. I added a feature that will attempt to
grab the books cover using its isbn and display it. This only worked for
about half the books so far so I might have to look into using a
different db for that.

It also adds some colors in a very ugly way. But I couldnt figure out
how to override default bootstrap colors so that is how it is for now.

And I cleared up some logic that was finding the book over and over
instead of just setting it to a variable.