Redo how most things work
I make a mistake and lost track of all the changes I made so this is a giant
ugly commit.

I changed the basic behavior of the app so that instead of the user creating a
book by filling out a form, they can search the book by name and it will use
the google books api gem to find the book and add it instead.